Refactoring to have globals splitted from opentx.h for easier includes
