+++ radio/src/gui/480x272/radio_spectrum_analyser.cpp
@@ -58,10 +58,7 @@ bool menuRadioSpectrumAnalyser(event_t event)
     lcdRefresh();
     if (isModulePXX2(g_moduleIdx)) {
       moduleState[g_moduleIdx].readModuleInformation(&reusableBuffer.moduleSetup.pxx2.moduleInformation, PXX2_HW_INFO_TX_ID, PXX2_HW_INFO_TX_ID);
-#if defined(ACCESS_LIB)
-      // the module will reset on mode switch, we need to reset the authentication counter
-      globalData.authenticationCount = 0;
-#endif
+      resetAccessAuthenticationCount();
     }
     else if (isModuleMultimodule(g_moduleIdx)) {
       if (reusableBuffer.spectrumAnalyser.moduleOFF)
