+++ radio/src/pulses/pulses.cpp
@@ -308,9 +308,7 @@ static void enablePulsesInternalModule(uint8_t protocol)
 #if defined(PXX2)
     case PROTOCOL_CHANNELS_PXX2_HIGHSPEED:
       intmoduleSerialStart(PXX2_HIGHSPEED_BAUDRATE, true, USART_Parity_No, USART_StopBits_1, USART_WordLength_8b);
-#if defined(HARDWARE_INTERNAL_MODULE) && defined(INTERNAL_MODULE_PXX2) && defined(ACCESS_LIB)
-      globalData.authenticationCount = 0;
-#endif
+      resetAccessAuthenticationCount();
       break;
 #endif
